ci: move common functions to common_ext library
